+++ src/Akeneo/Pim/Structure/Bundle/Resources/workspaces/settings-ui/src/pages/LocalesIndex.tsx
@@ -1,10 +1,9 @@
 import React, {FC, useEffect} from 'react';
-import {PimView} from '@akeneo-pim-community/legacy-bridge';
-import {PageContent, PageHeader, useRoute, useTranslate} from '@akeneo-pim-community/shared';
+import {Breadcrumb, getColor, Helper as BaseHelper} from 'akeneo-design-system';
+import {PageContent, PageHeader, useRoute, useTranslate, PimView} from '@akeneo-pim-community/shared';
 import {LocalesDataGrid} from '../components';
 import {useLocalesIndexState} from '../hooks';
 import styled from 'styled-components';
-import {Breadcrumb, getColor, Helper as BaseHelper} from 'akeneo-design-system';
 
 const Helper = styled(BaseHelper)`
   margin-bottom: 20px;