+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/public/hooks/use-measurement-families.ts
@@ -1,6 +1,6 @@
 import {useCallback, useEffect, useState} from 'react';
-import {MeasurementFamily} from 'akeneomeasure/model/measurement-family';
-import {baseFetcher} from 'akeneomeasure/shared/fetcher/base-fetcher';
+import {MeasurementFamily} from '../model/measurement-family';
+import {baseFetcher} from '../shared/fetcher/base-fetcher';
 import {useRoute} from '@akeneo-pim-community/shared';
 
 const useMeasurementFamilies = (): [MeasurementFamily[] | null, () => Promise<void>] => {